requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiolimiter==1.0.0
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
APScheduler==3.6.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
CacheControl==0.12.11
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect==1.2.0
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
elevenlabs==0.2.15
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.95.1
ffmpeg-python==0.2.0
filelock==3.12.0
firebase-admin==6.1.0
frozenlist==1.3.3
future==0.18.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.56.4
greenlet==2.0.2
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.172
llvmlite==0.40.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numba==0.57.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
orjson==3.8.11
packaging==23.1
pandas==2.0.2
parso==0.8.3
paypalrestsdk==1.13.1
pexpect==4.8.0
pickleshare==0.7.5
pinecone-client==2.2.1
playwright==1.34.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
pyee==9.0.4
Pygments==2.15.1
PyJWT==2.6.0
pymongo==4.3.3
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyshorteners==1.0.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==13.13
pytz==2023.3
PyYAML==6.0
redbird==0.7.1
regex==2023.5.4
requests==2.29.0
rich==13.3.5
rocketry==2.5.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
stack-data==0.6.2
starlette==0.26.1
stripe==5.4.0
structlog==22.3.0
sympy==1.12
tenacity==8.2.2
TikTokApi==5.2.2
tiktoken==0.3.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
twilio==8.2.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.15
uuid==1.30
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
yarl==1.9.2
